[Style] opt integration tests
